Created the functionality for the GUI to add and delete items from the database, using the database API.
Also created the GetAllItemGroups method for the database API.

+++ Kode/DatabaseAPI/DatabaseAPI/TableItem/ITableItem.cs
@@ -5,8 +5,8 @@ namespace DatabaseAPI.TableItem
 {
     public interface ITableItem
     {
-        void CreateItem(string name, int itemGroup);
-        void DeleteItem(int ID);
+        void CreateItem(string name, long itemGroup);
+        void DeleteItem(long ID);
         List<Item> SearchItems(string itemName);
     }
 }
